Move Analytics code from .usagetracker bundle
