+++ plugins/com.google.cloud.tools.eclipse.preferences.test/META-INF/MANIFEST.MF
@@ -10,10 +10,5 @@ Bundle-ActivationPolicy: lazy
 Fragment-Host: com.google.cloud.tools.eclipse.preferences
 Require-Bundle: org.hamcrest;bundle-version="1.1.0",
  org.junit;bundle-version="4.12.0"
-Import-Package: com.google.cloud.tools.eclipse.test.util,
- org.eclipse.core.resources,
- org.eclipse.jface.preference,
- org.mockito;provider=google;version="1.10.19",
- org.mockito.runners;provider=google;version="1.10.19",
- org.mockito.stubbing;provider=google;version="1.10.19",
- org.objenesis;provider=google;version="2.2.0"
+Import-Package: org.mockito;provider=google;version="1.10.19",
+ org.mockito.runners;provider=google;version="1.10.19"
