+++ plugins/com.google.cloud.tools.eclipse.preferences/plugin.properties
@@ -1,3 +1,2 @@
 extension.page.name=Google Cloud Tools
-analytics.optin.area.title=Usage Statistics
 extension.name=Preference Areas
