+++ plugins/com.google.cloud.tools.eclipse.preferences/plugin.xml
@@ -13,14 +13,5 @@
             class="com.google.cloud.tools.eclipse.preferences.areas.AreaBasedPreferencePage">
       </page>
    </extension>
-   <extension
-         point="com.google.cloud.tools.eclipse.preferences.areas">
-      <area
-            host="com.google.cloud.tools.eclipse.preferences.main"
-            title="%analytics.optin.area.title"
-            class="com.google.cloud.tools.eclipse.preferences.AnalyticsOptInArea"
-            preferences="configuration://com.google.cloud.tools.eclipse.usagetracker">
-      </area>
-   </extension>
 
 </plugin>
