+++ plugins/com.google.cloud.tools.eclipse.usagetracker/build.properties
@@ -1,5 +1,7 @@
 output.. = target/classes/
 bin.includes = META-INF/,\
+               plugin.xml,\
+               plugin.properties,\
                .
 source.. = src/,\
            generated-sources/
