+++ plugins/com.google.cloud.tools.eclipse.usagetracker/plugin.properties
@@ -1,0 +1,1 @@
+analytics.optin.area.title=Usage Statistics
